Conflicts:
	xbmc/cores/VideoRenderers/BaseRenderer.h
	xbmc/guilib/GUIControlGroup.cpp
	xbmc/utils/CharsetConverter.cpp
	xbmc/video/dialogs/GUIDialogAudioSubtitleSettings.cpp
	xbmc/video/dialogs/GUIDialogVideoSettings.cpp

+++ xbmc/pvr/timers/PVRTimers.cpp
@@ -18,24 +18,24 @@
  *
  */
 
-#include "FileItem.h"
+#include "PVRTimers.h"
+
+#include <cassert>
+#include <cstdlib>
+#include <utility>
+
 #include "dialogs/GUIDialogKaiToast.h"
 #include "dialogs/GUIDialogOK.h"
 #include "epg/EpgContainer.h"
+#include "FileItem.h"
+#include "pvr/addons/PVRClients.h"
+#include "pvr/channels/PVRChannelGroupsContainer.h"
+#include "pvr/PVRManager.h"
 #include "settings/Settings.h"
 #include "threads/SingleLock.h"
 #include "utils/log.h"
 #include "utils/StringUtils.h"
 #include "utils/Variant.h"
-
-#include "pvr/PVRManager.h"
-#include "pvr/addons/PVRClients.h"
-#include "pvr/channels/PVRChannelGroupsContainer.h"
-
-#include "PVRTimers.h"
-
-#include <assert.h>
-#include <cstdlib>
 
 using namespace PVR;
 using namespace EPG;
@@ -647,7 +647,7 @@ CPVRTimerInfoTagPtr CPVRTimers::GetByClient(int iClientId, unsigned int iClientT
     for (VecTimerInfoTag::const_iterator timerIt = it->second->begin(); timerIt != it->second->end(); ++timerIt)
     {
       if ((*timerIt)->m_iClientId == iClientId &&
-          (*timerIt)->m_iClientIndex == static_cast<int>(iClientTimerId))
+          (*timerIt)->m_iClientIndex == iClientTimerId)
         return *timerIt;
     }
   }
